Modifies package.json to work for deploying to heroku
If we're going to deploy to heroku, we can't have this development
setup thing with multiple servers. Instead we need to npm run
build. Also we need heroku to know to install all of the clientside
dependencies, so those are now in the serverside dependencies, which
is probably wrong but also works for now? **wince**